test(admin): share mobile e2e contract helpers
